+++ api/src/certification/enrolment/domain/services/sessions-import-validation-service.js
